Replace NSVisualEffectView with custom toolbar
Because theming will require custom coloring and some other stuff,
I've decided to replace NSVisualEffectView in navigation bar with
custom NSToolbar-lookalike implementation. Currently it's themed only
for white theme, because we hadn't settled on themes yet.

NSToolbar is not enough for our needs because we need to be able to
arrange items in a custom manner (like aligning with split view),
but I think we can embed NSToolbar (or many toolbars) inside this bar.